+++ Assets/Source/GuiHud/GUI.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: ead5327542ac61e438439c93130a9175
+guid: 06f5220653095964a83174e90aa0325d
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
