+++ Assets/Source/GuiHud/GUI/Elements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e948084e15469b947864dbad7d67a4c7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
