+++ Assets/Source/GuiHud/GUI/Elements/ProgressBar.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace KGUI.Elements
+{
+
+}
+public class ProgressBar
+{
+
+    private void Initialize()
+    {
+        
+    }
+
+    private void Update()
+    {
+        
+    }
+}
