+++ Assets/Source/GuiHud/GUI/Elements/ProgressBar.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e8d0fd2d7b551a4ca3273531dc3e5da
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
