+++ Assets/Source/GuiHud/GUI/GUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: defcbe98d73aa8745bdc01839dcae485
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
