+++ Assets/Source/GuiHud/GUI/PlayerStatusGUI/FoodBarUI.cs
@@ -1,37 +1,36 @@
+using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using Entitas;
 
-namespace KGUI
+namespace KGUI.PlayerStatus
 {
-    public class OxygenBarUI
+    public class FoodBarUI
     {
         // Init
         private static bool Init;
 
-        // Oxygen Bar Icon Position
+        // Food Bar Icon Position
         public Rect iconPosition = new Rect(7, 140, 60, -60);
 
-        // Oxygen Bar Icon Sprite
+        // Food Bar Icon Sprite
         Sprites.Sprite icon;
         Sprites.Sprite fill;
 
         // Image
-        public GameObject oxygenBar;
+        public GameObject foodBar;
         private GameObject iconCanvas;
 
-        Contexts EntitasContext;
-
-        public void Initialize(Contexts entitasContext)
+        public void Initialize(AgentEntity agentEntity)
         {
-            EntitasContext = entitasContext;
             // Set Width and Height
             int IconWidth = 19;
             int IconHeight = 19;
             Vector2Int iconPngSize = new Vector2Int(IconWidth, IconHeight);
 
             // Load image from file
-            var iconSheet = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\UserInterface\\Icons\\Oxygen\\hud_status_oxygen.png", IconWidth, IconHeight);
+            var iconSheet = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\UserInterface\\Icons\\Food\\hud_status_food.png", IconWidth, IconHeight);
 
             // Set Sprite ID from Sprite Atlas
             int iconID = GameState.SpriteAtlasManager.CopySpriteToAtlas(iconSheet, 0, 0, Enums.AtlasType.Particle);
@@ -79,71 +78,65 @@ namespace KGUI
                 TextureCoords = new Vector4(0, 0, 1, 1)
             };
 
-            // Oxygen Bar Initializon
-            iconCanvas = new GameObject("Oxygen Icon");
+            // Food Bar Initializon
+            iconCanvas = new GameObject("Food Icon");
             iconCanvas.transform.parent = GameObject.Find("Canvas").transform;
             iconCanvas.AddComponent<RectTransform>();
             iconCanvas.AddComponent<Image>();
 
-            // Add Components and setup agent object
+            // Add Components and setup game object
             Sprite iconBar = Sprite.Create(icon.Texture, new Rect(0.0f, 0.0f, IconWidth, IconHeight), new Vector2(0.5f, 0.5f));
             iconCanvas.GetComponent<Image>().sprite = iconBar;
 
             if (Camera.main.aspect >= 1.7f)
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 5.9f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 123.8f, 4.873917f);
             else if (Camera.main.aspect >= 1.5f)
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 9.6f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 125.4f, 4.873917f);
             else
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 75.3f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 193.4f, 4.873917f);
 
             iconCanvas.GetComponent<RectTransform>().localScale = new Vector3(0.6f, -0.6f, 0.5203559f);
 
-            // Oxygen Bar Initializon
-            oxygenBar = new GameObject("Oxygen Bar");
-            oxygenBar.transform.parent = iconCanvas.transform;
-            oxygenBar.AddComponent<RectTransform>();
-            oxygenBar.AddComponent<Image>();
+            // Food Bar Initializon
+            foodBar = new GameObject("Food Bar");
+            foodBar.transform.parent = iconCanvas.transform;
+            foodBar.AddComponent<RectTransform>();
+            foodBar.AddComponent<Image>();
 
-            // Add Components and setup agent object
+            // Add Components and setup game object
             Sprite bar = Sprite.Create(fill.Texture, new Rect(0.0f, 0.0f, FillWidth, FillHeight), new Vector2(0.5f, 0.5f));
 
-            oxygenBar.GetComponent<Image>().sprite = bar;
-            oxygenBar.GetComponent<Image>().raycastTarget = true;
-            oxygenBar.GetComponent<Image>().maskable = true;
-            oxygenBar.GetComponent<Image>().type = Image.Type.Filled;
-            oxygenBar.GetComponent<Image>().fillMethod = Image.FillMethod.Radial360;
-            oxygenBar.GetComponent<Image>().fillOrigin = 0;
-            IGroup<AgentEntity> Playerentities =
-            EntitasContext.agent.GetGroup(AgentMatcher.AgentStats);
-            foreach (var entity in Playerentities)
+            foodBar.GetComponent<Image>().sprite = bar;
+            foodBar.GetComponent<Image>().raycastTarget = true;
+            foodBar.GetComponent<Image>().maskable = true;
+            foodBar.GetComponent<Image>().type = Image.Type.Filled;
+            foodBar.GetComponent<Image>().fillMethod = Image.FillMethod.Radial360;
+            foodBar.GetComponent<Image>().fillOrigin = 0;
+            float foodValue = agentEntity.agentStats.Food;
+            if (foodValue <= 0)
             {
-                oxygenBar.GetComponent<Image>().fillAmount = entity.agentStats.Oxygen / 100;
+                foodValue = 0;
             }
-            oxygenBar.GetComponent<Image>().fillClockwise = true;
-            oxygenBar.GetComponent<RectTransform>().localPosition = new Vector3(-0.4f, -0.1f, 4.873917f);
-
-            oxygenBar.GetComponent<RectTransform>().localScale = new Vector3(0.8566527f, 0.8566527f, 0.3714702f);
+            foodBar.GetComponent<Image>().fillAmount = foodValue / 100;
+            foodBar.GetComponent<Image>().fillClockwise = true;
+            foodBar.GetComponent<RectTransform>().localPosition = new Vector3(-0.4f, -0.1f, 4.873917f);
+            foodBar.GetComponent<RectTransform>().localScale = new Vector3(0.8566527f, 0.8566527f, 0.3714702f);
 
             Init = true;
         }
 
-        public void Update()
+        public void Update(AgentEntity agentEntity)
         {
             if (Init)
             {
-                IGroup<AgentEntity> Playerentities =
-                EntitasContext.agent.GetGroup(AgentMatcher.AgentStats);
-                foreach (var entity in Playerentities)
-                {
-                    oxygenBar.GetComponent<Image>().fillAmount = entity.agentStats.Oxygen / 100;
-                }
+                foodBar.GetComponent<Image>().fillAmount = agentEntity.agentStats.Food / 100;
 
                 if (Camera.main.aspect >= 1.7f)
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 5.9f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 123.8f, 4.873917f);
                 else if (Camera.main.aspect >= 1.5f)
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 9.6f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 125.4f, 4.873917f);
                 else
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 75.3f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 193.4f, 4.873917f);
             }
         }
     }
