+++ Assets/Source/GuiHud/GUI/PlayerStatusGUI/FoodBarUI.cs.meta
