+++ Assets/Source/GuiHud/GUI/PlayerStatusGUI/FuelBarUI.cs.meta
