+++ Assets/Source/GuiHud/GUI/PlayerStatusGUI/WaterBarUI.cs
@@ -1,38 +1,34 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 using Entitas;
 
-namespace KGUI
+namespace KGUI.PlayerStatus
 {
-    public class FoodBarUI
+    public class WaterBarUI
     {
         // Init
         private static bool Init;
 
-        // Food Bar Icon Position
+        // Water Bar Icon Position
         public Rect iconPosition = new Rect(7, 140, 60, -60);
 
-        // Food Bar Icon Sprite
+        // Water Bar Icon Sprite
         Sprites.Sprite icon;
         Sprites.Sprite fill;
 
         // Image
-        public GameObject foodBar;
+        public GameObject waterBar;
         private GameObject iconCanvas;
-        Contexts EntitasContext;
 
-        public void Initialize(Contexts entitasContext)
+        public void Initialize(AgentEntity agentEntity)
         {
-            EntitasContext = entitasContext;
             // Set Width and Height
             int IconWidth = 19;
             int IconHeight = 19;
             Vector2Int iconPngSize = new Vector2Int(IconWidth, IconHeight);
 
             // Load image from file
-            var iconSheet = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\UserInterface\\Icons\\Food\\hud_status_food.png", IconWidth, IconHeight);
+            var iconSheet = GameState.SpriteLoader.GetSpriteSheetID("Assets\\StreamingAssets\\UserInterface\\Icons\\Water\\hud_status_water.png", IconWidth, IconHeight);
 
             // Set Sprite ID from Sprite Atlas
             int iconID = GameState.SpriteAtlasManager.CopySpriteToAtlas(iconSheet, 0, 0, Enums.AtlasType.Particle);
@@ -80,8 +76,8 @@ namespace KGUI
                 TextureCoords = new Vector4(0, 0, 1, 1)
             };
 
-            // Food Bar Initializon
-            iconCanvas = new GameObject("Food Icon");
+            // Water Bar Initializon
+            iconCanvas = new GameObject("Water Icon");
             iconCanvas.transform.parent = GameObject.Find("Canvas").transform;
             iconCanvas.AddComponent<RectTransform>();
             iconCanvas.AddComponent<Image>();
@@ -91,61 +87,54 @@ namespace KGUI
             iconCanvas.GetComponent<Image>().sprite = iconBar;
 
             if (Camera.main.aspect >= 1.7f)
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 123.8f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 64.9f, 4.873917f);
             else if (Camera.main.aspect >= 1.5f)
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 125.4f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 67f, 4.873917f);
             else
-                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 193.4f, 4.873917f);
+                iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 134.2f, 4.873917f);
+
 
             iconCanvas.GetComponent<RectTransform>().localScale = new Vector3(0.6f, -0.6f, 0.5203559f);
 
-            // Food Bar Initializon
-            foodBar = new GameObject("Food Bar");
-            foodBar.transform.parent = iconCanvas.transform;
-            foodBar.AddComponent<RectTransform>();
-            foodBar.AddComponent<Image>();
+            // Water Bar Initializon
+            waterBar = new GameObject("Water Bar");
+            waterBar.transform.parent = iconCanvas.transform;
+            waterBar.AddComponent<RectTransform>();
+            waterBar.AddComponent<Image>();
 
             // Add Components and setup game object
             Sprite bar = Sprite.Create(fill.Texture, new Rect(0.0f, 0.0f, FillWidth, FillHeight), new Vector2(0.5f, 0.5f));
 
-            foodBar.GetComponent<Image>().sprite = bar;
-            foodBar.GetComponent<Image>().raycastTarget = true;
-            foodBar.GetComponent<Image>().maskable = true;
-            foodBar.GetComponent<Image>().type = Image.Type.Filled;
-            foodBar.GetComponent<Image>().fillMethod = Image.FillMethod.Radial360;
-            foodBar.GetComponent<Image>().fillOrigin = 0;
-            IGroup<AgentEntity> Playerentities =
-            EntitasContext.agent.GetGroup(AgentMatcher.AgentStats);
-            foreach (var entity in Playerentities)
-            {
-                foodBar.GetComponent<Image>().fillAmount = entity.agentStats.Food / 100;
-            }
-            foodBar.GetComponent<Image>().fillClockwise = true;
-            foodBar.GetComponent<RectTransform>().localPosition = new Vector3(-0.4f, -0.1f, 4.873917f);
+            waterBar.GetComponent<Image>().sprite = bar;
+            waterBar.GetComponent<Image>().raycastTarget = true;
+            waterBar.GetComponent<Image>().maskable = true;
+            waterBar.GetComponent<Image>().type = Image.Type.Filled;
+            waterBar.GetComponent<Image>().fillMethod = Image.FillMethod.Radial360;
+            waterBar.GetComponent<Image>().fillOrigin = 0;
+            waterBar.GetComponent<Image>().fillAmount = agentEntity.agentStats.Water / 100;
+            waterBar.GetComponent<Image>().fillClockwise = true;
 
+            waterBar.GetComponent<RectTransform>().localPosition = new Vector3(-0.4f, -0.1f, 4.873917f);
 
-            foodBar.GetComponent<RectTransform>().localScale = new Vector3(0.8566527f, 0.8566527f, 0.3714702f);
+            waterBar.GetComponent<RectTransform>().localScale = new Vector3(0.8566527f, 0.8566527f, 0.3714702f);
 
             Init = true;
         }
 
-        public void Update()
+        public void Update(AgentEntity agentEntity)
         {
             if (Init)
             {
                 IGroup<AgentEntity> Playerentities =
-                EntitasContext.agent.GetGroup(AgentMatcher.AgentStats);
-                foreach (var entity in Playerentities)
-                {
-                    foodBar.GetComponent<Image>().fillAmount = entity.agentStats.Food / 100;
-                }
+                Contexts.sharedInstance.agent.GetGroup(AgentMatcher.AgentStats);
+                waterBar.GetComponent<Image>().fillAmount = agentEntity.agentStats.Water / 100;
 
                 if (Camera.main.aspect >= 1.7f)
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 123.8f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-377.3f, 64.9f, 4.873917f);
                 else if (Camera.main.aspect >= 1.5f)
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 125.4f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-335.6f, 67f, 4.873917f);
                 else
-                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 193.4f, 4.873917f);
+                    iconCanvas.GetComponent<RectTransform>().localPosition = new Vector3(-363.8f, 134.2f, 4.873917f);
             }
         }
     }
