+++ Assets/Source/GuiHud/GUI/PlayerStatusGUI/WaterBarUI.cs.meta
