+++ Assets/Source/GuiHud/HUD.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 17d4393f2e4ceed42b358f75bbad5c55
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
